Changing to using the timestamp to compare
Our CFBundleVersion is updated at runtime so it isn’t available on
HockeyApp. To work around this we can parse the CFBundleVersion at
runtime and compare it to the version timestamp .

We expect to find 6 numbers in a row at some point after ‘(‘.

The numbers are expected to be formatted yyMMdd - with the two digit
minimum year being Jan 1 2000.